+++ src/flip7/envs/__init__.py
@@ -1,1 +1,41 @@
 """RL environment adapters for the Flip 7 engine."""
+
+from flip7.envs.aec import Flip7AECEnv
+from flip7.envs.encodings import (
+    HIT_INDEX,
+    STAY_INDEX,
+    TARGET_OFFSET,
+    action_mask,
+    action_space_size,
+    agent_name,
+    decode_action,
+    encode_action,
+    player_id_from_agent,
+)
+from flip7.envs.observations import (
+    ObservationFamily,
+    encode_observation,
+    observation_size,
+)
+from flip7.envs.rewards import RewardMode, rewards_from_result
+from flip7.envs.vs_opponents import Flip7VsOpponentsEnv, RandomLegalPolicy
+
+__all__ = [
+    "HIT_INDEX",
+    "STAY_INDEX",
+    "TARGET_OFFSET",
+    "Flip7AECEnv",
+    "Flip7VsOpponentsEnv",
+    "ObservationFamily",
+    "RandomLegalPolicy",
+    "RewardMode",
+    "action_mask",
+    "action_space_size",
+    "agent_name",
+    "decode_action",
+    "encode_action",
+    "encode_observation",
+    "observation_size",
+    "player_id_from_agent",
+    "rewards_from_result",
+]